+++ Catalog.md
@@ -4,23 +4,21 @@
     │   │   ├── cont
     │   │   │   ├── ab
     │   │   │   │   ├── builder.rs
-    │   │   │   │   ├── chunking.rs
-    │   │   │   │   ├── container.rs
+    │   │   │   │   ├── congruent.rs
+    │   │   │   │   ├── eclectic.rs
     │   │   │   │   ├── mod.rs
     │   │   │   │   ├── private.rs
     │   │   │   │   ├── purport.rs
-    │   │   │   │   ├── seq.rs
-    │   │   │   │   └── sheaf.rs
-    │   │   │   ├── array.rs
+    │   │   │   │   └── seq.rs
     │   │   │   ├── batch.rs
     │   │   │   ├── bundle.rs
     │   │   │   ├── cvt.rs
-    │   │   │   ├── grid.rs
+    │   │   │   ├── ext.rs
+    │   │   │   ├── macros.rs
     │   │   │   ├── mod.rs
     │   │   │   ├── nullopt.rs
     │   │   │   ├── parcel.rs
-    │   │   │   ├── table.rs
-    │   │   │   └── vector.rs
+    │   │   │   └── table.rs
     │   │   ├── io
     │   │   │   ├── ab
     │   │   │   │   └── mod.rs
@@ -30,6 +28,7 @@
     │   │   │   ├── mod.rs
     │   │   │   ├── parquet.rs
     │   │   │   └── sql.rs
+    │   │   ├── ctor.rs
     │   │   ├── error.rs
     │   │   ├── lib.rs
     │   │   ├── macros.rs
